requirements.txt
Flask==2.0.2
requests==2.27.1
beautifulsoup4==4.10.0
bs4==0.0.1
gunicorn==20.1.0
pytest-shutil==1.7.0
pymongo==4.0.1
dnspython==2.2.0
Pillow==9.0.0